+++ lib/braintree/customer.rb
@@ -101,14 +101,6 @@ module Braintree
       return_object_or_raise(:transaction) { credit(transaction_attributes) }
     end
 
-    # Deprecated. Use Braintree::Customer.default_payment_method
-    #
-    # Returns the customer's default credit card.
-    def default_credit_card
-      warn "[DEPRECATED] Customer#default_credit_card is deprecated. Please use Customer#default_payment_method"
-      @credit_cards.find { |credit_card| credit_card.default? }
-    end
-
     # Returns the customer's default payment method.
     def default_payment_method
       payment_methods.find { |payment_instrument| payment_instrument.default? }
@@ -140,37 +132,9 @@ module Braintree
       "#<#{self.class} #{nice_attributes.join(', ')}>"
     end
 
-    # Deprecated. Use Braintree::Customer.sale
-    def sale(transaction_attributes)
-      warn "[DEPRECATED] sale as an instance method is deprecated. Please use Customer.sale"
-      @gateway.transaction.sale(transaction_attributes.merge(:customer_id => id))
-    end
-
-    # Deprecated. Use Braintree::Customer.sale!
-    def sale!(transaction_attributes)
-      warn "[DEPRECATED] sale! as an instance method is deprecated. Please use Customer.sale!"
-      return_object_or_raise(:transaction) { sale(transaction_attributes) }
-    end
-
     # Returns a ResourceCollection of transactions for the customer.
     def transactions(options = {})
       @gateway.customer.transactions(id, options)
-    end
-
-    # Deprecated. Use Braintree::Customer.update
-    def update(attributes)
-      warn "[DEPRECATED] update as an instance method is deprecated. Please use Customer.update"
-      result = @gateway.customer.update(id, attributes)
-      if result.success?
-        copy_instance_variables_from_object result.customer
-      end
-      result
-    end
-
-    # Deprecated. Use Braintree::Customer.update!
-    def update!(attributes)
-      warn "[DEPRECATED] update! as an instance method is deprecated. Please use Customer.update!"
-      return_object_or_raise(:customer) { update(attributes) }
     end
 
     class << self