+++ lib/braintree/dispute.rb
@@ -121,11 +121,5 @@ module Braintree
         Braintree::Dispute::HistoryEvent.new(event)
       end unless status_history.nil?
     end
-
-    def forwarded_comments
-      # NEXT_MAJOR_VERSION delete this method since it never returned anything anyway.
-      warn "[DEPRECATED] #forwarded_comments is deprecated. Please use #processor_comments"
-      processor_comments
-    end
   end
 end