+++ lib/braintree/error_codes.rb
@@ -520,7 +520,6 @@ module Braintree
       DescriptionIsTooLong = "95803"
       DiscountAmountFormatIsInvalid = "95804"
       DiscountAmountIsTooLarge = "95805"
-      DiscountAmountMustBeGreaterThanZero = "95806" # Deprecated as the amount may be zero. Use DiscountAmountCannotBeNegative.
       DiscountAmountCannotBeNegative = "95806"
       KindIsInvalid = "95807"
       KindIsRequired = "95808"
@@ -541,7 +540,6 @@ module Braintree
       UnitOfMeasureIsTooLong = "95821"
       UnitTaxAmountFormatIsInvalid = "95824"
       UnitTaxAmountIsTooLarge = "95825"
-      UnitTaxAmountMustBeGreaterThanZero = "95826" # Deprecated as the amount may be zero. Use UnitTaxAmountCannotBeNegative.
       UnitTaxAmountCannotBeNegative = "95826"
       TaxAmountFormatIsInvalid = "95827"
       TaxAmountIsTooLarge = "95828"